+++ client/containers/Login/LoginWrap.js
@@ -9,7 +9,7 @@ const TabPane = Tabs.TabPane;
 
 @connect(state => ({
   loginWrapActiveKey: state.user.loginWrapActiveKey,
-  canRegister: state.canRegister
+  canRegister: state.user.canRegister
 }))
 export default class LoginWrap extends Component {
   constructor(props) {
@@ -23,7 +23,7 @@ export default class LoginWrap extends Component {
   };
 
   render() {
-    const { loginWrapActiveKey } = this.props;
+    const { loginWrapActiveKey, canRegister } = this.props;
     {/** show only login when register is disabled */}
     return (
       <Tabs
@@ -34,7 +34,7 @@ export default class LoginWrap extends Component {
         <TabPane tab="登录" key="1">
           <LoginForm />
         </TabPane>
-        <TabPane tab="注册" key="2">
+        <TabPane tab="注册" key="2" disabled={!canRegister}>
           <RegForm />
         </TabPane>
       </Tabs>