+++ server/controllers/open.js
@@ -233,8 +233,8 @@ class openController extends baseController {
     };
 
     if (ctx.params.email === true && reportsResult.message.failedNum !== 0) {
-      let autoTestUrl = `http://${
-        ctx.request.host
+      let autoTestUrl = `${
+        ctx.request.origin
       }/api/open/run_auto_test?id=${id}&token=${token}&mode=${ctx.params.mode}`;
       yapi.commons.sendNotice(projectId, {
         title: `YApi自动化测试报告`,